Day 25 - Python Discord bot/requirements.txt
requests==2.25.1
discord.py==1.7.3
beautifulsoup4==4.9.3
discord==1.7.3
Pillow==8.2.0
python-dotenv==0.18.0
